Everything works

+++ frontend/src/app/login/login.component.ts
@@ -3,7 +3,6 @@ import { Router } from '@angular/router';
 import { LoginService } from './login.service';
 import { NgForm } from '@angular/forms';
 import { AxiosService } from '../axios.service';
-import { AuthenticationService } from '../authentication.service'; // Import AuthenticationService
 import { UsersService } from '../users/users.service'; // Import UsersService
 
 @Component({
@@ -17,36 +16,47 @@ export class LoginComponent implements OnInit {
   username: string = '';
   password: string = '';
   passwordVisible = false;
-  showPasswordMismatchWarning = false;
+  somethingWrong = false;
 
   constructor(private router: Router,
-              private loginService: LoginService,
               private axiosService: AxiosService,
-              private authService: AuthenticationService  
+              private userService: UsersService
             ) { }
 
   ngOnInit() {
-    window.localStorage.clear();
+    if (typeof window !== "undefined") {  
+      window.localStorage.clear();
+    }
   }
 
   
-onSubmitLogin(loginForm: NgForm) {
-  const value = loginForm.value;
-  console.log('Form Value:', value);
-
-  this.loginService.verifyCredentials(value.username, value.password).subscribe(
-    (response) => {
-      this.authService.login(value.username, value.password);
-      this.router.navigate(['/welcome']);
-    },
-    (error) => {
-      console.log('Error:', error);
-      if (error.status === 401) {
-        this.showPasswordMismatchWarning = true;
+  onSubmitLogin(loginForm: NgForm) {
+    const value = loginForm.value; 
+    this.axiosService.request(
+      "POST",
+      "/login",
+      {
+        userId: value.username, 
+        userPassword: value.password, 
       }
-    }
-  );
-}
+    ).then(
+      response => {
+        console.log('Response:', response.data.token);
+        this.axiosService.setAuthToken(response.data.token);
+        
+        const username =value.username;
+        this.userService.setUsername(username);        
+
+        this.router.navigate(['/welcome']);
+      }
+    ).catch(
+      error => {
+        if (error.response.status === 401) {
+          this.somethingWrong = true;
+        }
+      }
+    );
+  }
 
   togglePasswordVisibility() {
     this.passwordVisible = !this.passwordVisible;
